pull each assignment into its own file
