+++ Assignments/Assignments.csproj
@@ -43,6 +43,9 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="DayAssignment.cs" />
+    <Compile Include="EasterSundayAssignment.cs" />
+    <Compile Include="HailstoneAssignment.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
